+++ Telephone/CallHistoryViewEventTarget.swift
@@ -21,25 +21,22 @@ import UseCases
 @MainActor
 final class CallHistoryViewEventTarget: NSObject {
     private let recordsGet: UseCase
-    private let purchaseCheck: UseCase
     private let recordRemoveAll: UseCase
     private let recordRemove: CallHistoryRecordRemoveUseCaseFactory
     private let callMake: CallHistoryCallMakeUseCaseFactory
 
     init(recordsGet: UseCase,
-         purchaseCheck: UseCase,
          recordRemoveAll: UseCase,
          recordRemove: CallHistoryRecordRemoveUseCaseFactory,
          callMake: CallHistoryCallMakeUseCaseFactory) {
         self.recordsGet = recordsGet
-        self.purchaseCheck = purchaseCheck
         self.recordRemoveAll = recordRemoveAll
         self.recordRemove = recordRemove
         self.callMake = callMake
     }
 
     func shouldReloadData() {
-        executeRecordGetAndPurchaseCheck()
+        executeRecordGet()
     }
 
     func shouldRemoveAllRecords() {
@@ -54,40 +51,19 @@ final class CallHistoryViewEventTarget: NSObject {
         recordRemove.make(identifier: identifier).execute()
     }
 
-    private func executeRecordGetAndPurchaseCheck() {
+    private func executeRecordGet() {
         recordsGet.execute()
-        purchaseCheck.execute()
     }
 }
 
 nonisolated extension CallHistoryViewEventTarget: CallHistoryEventTarget {
     func didUpdate(_ history: CallHistory) {
-        Task {
-            await executeRecordGetAndPurchaseCheck()
-        }
+        Task { await executeRecordGet() }
     }
-}
-
-extension CallHistoryViewEventTarget: StoreEventTarget {
-    func didPurchase() {
-        executeRecordGetAndPurchaseCheck()
-    }
-
-    func didRestorePurchases() {
-        executeRecordGetAndPurchaseCheck()
-    }
-
-    func didStartPurchasingProduct(withIdentifier identifier: String) {}
-    func didFailPurchasing(error: String) {}
-    func didCancelPurchasing() {}
-    func didFailRestoringPurchases(error: String) {}
-    func didCancelRestoringPurchases() {}
 }
 
 nonisolated extension CallHistoryViewEventTarget: DayChangeEventTarget {
     func dayDidChange() {
-        Task {
-            await executeRecordGetAndPurchaseCheck()
-        }
+        Task { await executeRecordGet() }
     }
 }