+++ Telephone/CallHistoryViewEventTargetFactory.swift
@@ -23,33 +23,27 @@ final class CallHistoryViewEventTargetFactory {
     private let histories: CallHistories
     private let index: ContactMatchingIndex
     private let settings: ContactMatchingSettings
-    private let receipt: Receipt
     private let dateFormatter: DateFormatter
     private let durationFormatter: DateComponentsFormatter
-    private let storeEventTargets: StoreEventTargets
     private let dayChangeEventTargets: DayChangeEventTargets
 
     init(
         histories: CallHistories,
         index: ContactMatchingIndex,
         settings: ContactMatchingSettings,
-        receipt: Receipt,
         dateFormatter: DateFormatter,
         durationFormatter: DateComponentsFormatter,
-        storeEventTargets: StoreEventTargets,
         dayChangeEventTargets: DayChangeEventTargets,
         ) {
         self.histories = histories
         self.index = index
         self.settings = settings
-        self.receipt = receipt
         self.dateFormatter = dateFormatter
         self.durationFormatter = durationFormatter
-        self.storeEventTargets = storeEventTargets
         self.dayChangeEventTargets = dayChangeEventTargets
     }
 
-    func make(account: Account, view: CallHistoryView, purchaseCheck: UseCase) async -> CallHistoryViewEventTarget {
+    func make(account: Account, view: CallHistoryView) async -> CallHistoryViewEventTarget {
         let history = await histories.history(withUUID: account.uuid)
         let factory = FallingBackMatchedContactFactory(
             matching: IndexedContactMatching(
@@ -63,21 +57,16 @@ final class CallHistoryViewEventTargetFactory {
                 history: history,
                 output: ContactCallHistoryRecordGetAllUseCase(
                     factory: factory,
-                    output: ReceiptValidatingContactCallHistoryRecordGetAllUseCaseOutput(
-                        origin: CallHistoryViewPresenter(
-                            view: view, dateFormatter: dateFormatter, durationFormatter: durationFormatter
-                        ),
-                        receipt: receipt
+                    output: CallHistoryViewPresenter(
+                        view: view, dateFormatter: dateFormatter, durationFormatter: durationFormatter
                     )
                 )
             ),
-            purchaseCheck: purchaseCheck,
             recordRemoveAll: CallHistoryRecordRemoveAllUseCase(history: history),
             recordRemove: DefaultCallHistoryRecordRemoveUseCaseFactory(history: history),
             callMake: DefaultCallHistoryCallMakeUseCaseFactory(account: account, history: history, factory: factory)
         )
         await history.updateTarget(WeakCallHistoryEventTarget(origin: result))
-        storeEventTargets.add(WeakStoreEventTarget(origin: result))
         dayChangeEventTargets.add(WeakDayChangeEventTarget(origin: result))
         return result
     }